+++ openldap/CHANGELOG.md
@@ -1,6 +1,7 @@
 1.21
 
 * Version bump for new base image
+* Adjust for php-fpm/php_fpm in RC service
 
 ---